move details from workflow list to pages folder
