+++ packages/frinx-workflow-ui/src/pages/executed-workflow-detail/executed-workflow-detail-tabs/input-output-tab.tsx
